Prevent flickering in live preview thumbnails
Previously, we completely recreated the progress tooltip contents
whenever the cursor moved or we updated the stream progress. That's
both inefficient and causes flickering when displaying the live preview
thumbnails. Instead, we should detect when we're already showing a
progress tooltip and selectively update the contents, which is both more
efficient and prevents live preview thumbnail flickering.