Improve TVDb interface
Make a wrapper class for Series information, and if we fail to get
information about a particular episode for active streams, fall back
to the series information.

Also fix an issue where we would report that a season is incomplete if
the number of available episodes was greater than the total episodes
reported (e.g. a two-part finale that TVDb lists as a single episode).