Check all parts for potential preview thumbnails
A media item may have multiple versions, and it's not guaranteed that
all versions have generated thumbnails. If we fail to get a thumbnail
for one part, try all the other parts before giving up.

This change also requires modifications to the tooltip infrastrucure to
allow actual elements to be passed in as the tooltip opposed to just
strings. If only strings are allowed, we hit race conditions during the
element->string->element transition that may cause us to disregard
a media part before actually testing it, as we load the image source
twice: once during the initial construction, and again after converting
it to a string and subsequently rebuilding it as a DOM element.