Add ViewStream requests to activities
When the logic to add ViewStream requests changed, a bug was introduced
that prevented us from adding the request to the activity table unless
it had a comment attached, in which case the comment would not be added
to the request table. This change ensures we always attempt to add
activities for both the ViewStream request and the attached comment, if
provided.