Overhaul when notifications get sent
The logic to send email/text notifications was a bit broken. Previously,
comment notifications would only be sent if the person adding the
comment was an admin, or wasn't the creator of the request.

Change the logic to always send notifications to admins and the request
owner as long as the recipient isn't the creator.