Various a11y, perf, compat, security improvements
The canary build of Edge Chromium has an "Issues" tab that reports
problems with a particular page. This change addresses many of those
issues. I don't agree with the severity of many of Edge's complaints,
but they don't hurt to address either, so I did so wherever it wasn't a
complete pain to implement.

This also included some apache/php configuration changes that are not
captured in this repo, including changing default MIME types, better
default cache control, and turning off PHP/Apache server headers.