Mark HtmlStyle as a block element
The logic that inserts div/breaks doesn't expect inline elements to
contain 2+ newlines in a row, so when they do divs can incorrectly
splice the element, resulting in a bad parse. Mark HtmlStyle as a block
element, and adjust some tests accordingly.

Note that the entire _parseNewlines routine is overly convoluted,
fragile, and inconsistent, so some tests had to be adjusted to account
for that. Ideally the routine should be revamped, but the current state
of things is already the result of an overhaul of an even worse version.