Add volatile flag for Markdown elements
Some elements are volatile in the sense that even if the element itself
is not modified, its value can change if different element is modified
(e.g. ReferenceUrls and HtmlSpans with a class="" attribute). Introduce
a volatile flag that tells the converter to ignore cached values if the
current run tree contains a volatile element.