Rework markdown block state
In another effort to improve the auto-break/div logic, introduce a
'semi-block' state that will be treated as an inline element when
determining whether to insert breaks/divs in surrounding elements, and
as a block element when parsing newlines within the element.

In the end, things are slightly better (more texts follow the expected
pattern of all inline elements being in a div), but the overall logic
is more convoluted than ever.

Along the way, additional tangential changes were made:
  * Be more aggressive about removing inline elements from the cache
  * Remove extra newlines when trimming text between inline elements
  * Add text-transform as an allowed CSS attribute
  * Remove dead code