Wait to parse table cells until style is finalized
With my current horrible markdown style system that collapses and
inlines all styles, when table cells are computed in the middle of
parse, any styles defined after the table won't apply to the cells. To
get around this, wait to parse individual cells until after we've
completed the initial text parse. This also means creating real runs for
table elements (tr/td/etc) instead of building it manually within the
Table run's transform method.

This change also involves reworking how inlineOnly is conveyed to the
parser. Instead of an explicit parameter, create a thin class on top of
Markdown that sets inlineOnly automatically, and also takes in a style
object from the parent parser so we don't have to dig into the internals
of the parser to set the parent style.