Prevent extra breaks from being added in Markdown
In the never-ending story of errors in parseNewlines, if we have:

# A
B
C
# D

# E

and add another newline between #D and #E, another break will be added
between B and C. This is happening because when parsing newlines from a
cached string, we come across the same newline that we found during the
last parse, but still insert a new break. We should check whether the
place we're inserting the break was previously cached, and don't insert
the break if that's the case.