Fix immediate markdown list nesting.
If we have something like this:

1. * A
  * B

We should have an outer list with a single list item, a list with two
list items. Instead, '* B' is nested inside of '* A'. This happens
because the nest level logic doesn't account for the previous list
definition on the same level, so we thing '* A' has a nest level of 0,
and '* B' has a nest level of 1. Tweak things so we recognize when we're
immediately inside of a list and base our nest level on our parent list.