Indicate buffering sessions in session list
Previously anything that wasn't paused was considered playing, which
marked buffering items as playing. Instead of returning a boolean
"paused" status, return the full status so we can differentiate between
playing, buffering, and paused. And with that new state information,
use a loading icon instead of the play icon if a session is buffering.