Use maxOffsetAvailable for transcoder progress
TranscodeSession's progress field indicates its progress relative to the
point it started transcoding from. So if a movie starts 50% of the way
through and the transcoder has transcoded from 50% to 55%, it will
report 5%. When displaying progress bars though, we're more interested
in the maximum point of the transcoder, so instead use a combination of
the duration of the item and maxOffsetAvailable to calculate the "real"
progress of the transcoder.