Adjust mobile styling
For some phones, bringing up the keyboard convinces the browser that the
the orientation is now landscape, which results in incorrect styling
when making a new request or otherwise typing something in to the site.
The workaround is to remove explicit references to landscape and
portrait, relying just on pixel widths to determine mobile vs non-mobile
styling.